\name{vmsR}
\alias{vmsR}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
vmsR
}
\description{
It is a package that automatically processes and analyzes VMS data. Its functionalities include: Analyze raw VMS data; cleaning of erroneous data; identification of fishing trip trajectories using identification algorithms; Identify fishing hauls using artificial neural networks. In addition, it allows generating refined fishing effort measurements and spatial indicators of the fleet.
}

\references{
\url{https://github.com/mawp/spict/}
}
\author{
Pablo Marin A. \email{boosterpablo@gmail.com}
}
